Use "Players in View Angle" to determine when to shoot
When combined with "Start Facing" on a target, using "Ray Cast Hit Player" with a hitscan weapon results in 100% weapon accuracy. Obviously this is not realistic -- unless you are simply trying to create the perfect aimbot.

Using "Players in View Angle" with a small view angle (currently 5%) instead results in roughly 76% weapon accuracy (over a 10-minute period). This was determined using a 1v1 skirmish with "Hard Cassidy" on Team 1 and the custom Cassidy bot on Team 2. Note that the custom Cassidy bot does not currently move or use abilities (e.g. dodge using combat roll) so the view angle will need to be adjusted to achieve the desired weapon accuracy for each bot difficulty level after adding movement and abilities.