Fix issue where bot shoots at enemy that has not yet spawned
When using "Players in View Angle", Blizzard apparently expects us to explicitly exclude players who have not yet spawned...kind of lame, but it is what it is.

This issue is most noticeable when initially starting a skirmish with a human player on Team 1 -- i.e. before selecting a hero, the custom bot on Team 2 starts shooting immediately.